If status 200 and non-empty response, update Octavia LB object to ACTIVE
Resolves #6
Should the response HTTP status NOT be 200, an exception is thrown and no
response returned. Therefore the status is really only being updated to
ACTIVE/ONLINE if (and only if) response status was 200 and response body not
empty. AS3 always echoes the POSTed declaration in the response, so it's
reasonable to rely on it to be non-empty.