added multi-agent and scheduling capability
loadbalancers will be automatically scheduled to the
device with the lowest amount of listeners.
(amphorae with loadbalancer_id == None and
status == AMPHORAE_READY are BigIP devices)

After scheduling, the agent creates an amphora binding with
amphora.id == loadbalancer.id.

In case of failed scheduling, the reconciliation loop
of any working agent will pick up the pending loadbalancer
and associate it.

fixes #4