as3restclient: Add Prometheus metric for HTTP response status codes
The status code itself is a metric label, as well as the HTTP method
that was used.

Since the HTTP method has to be written manually when manipulating the
metric, it has to be manually changed when changing the request
method. This is a source of errors. I will address this in a separate
pull request.