Status manager: Set device entry to BOOTING if it comes back online
If a device was offline (entry in amphora table has
status=='ALLOCATED'), but is available now (it came back online), set
status to 'BOOTING'. This indicates that the device is back online but
needs a full sync.

It is the responsibility of the syncing mechanism to set the status to
'READY' again.