Use a new instance of AS3RestClient instead of modifying existing one
I previously relied on the acquired lock 'tenant_refresh' to keep
other functions from using the wrong BigIp within the time window it
was temporarily changed. But that is error-prone (e. g. if acquiring
the lock would be forgotten in a future code change).

Therefore use a second instance of AS3RestClient.
This however requires self._refresh to have an additional optional
parameter for the BigIp.