[status_manager] reworked status handling, status_update done in refresh()
earlier implementation just updates status of every loadbalancer
individually, although all changes in the tenant (loadbalancer in the
same network) have been pushed to the device through refresh().

This caused uneccesary updates to the same tenant with the same
declaration for every loadbalancer updated.

reworked status_update so it updates the status of all objects in the
same tenant and centralized status-updates.