[controller_worker] Member creation/update: Check for invalid IP
An IP address of a member is invalid if it collides with the IP
address of any VIP in the same network.

For now, the error is reported via the provisioning_status, but that
should be changed in cooperation with the GUI so as to not confuse the
customer with members not being created.