[status_manager] Fix: LBs with erroneous member stay PENDING
status_manager._update_status_to_octavia assumed that status['loadbalancers']
exists. When we only update the status of e. g. a member, it however does not.
This raises an exception.
This in turn aborts the tenant_update operation and the LB stays in PENDING*
state, blocking all LBs in the network.