At worker startup, register it to its assigned AZ
Also: Clean up ControllerWorker.__init__ a bit

We don't use AZs in the workers like the Amphora driver does for scheduling
amphoras. AZ-based scheduling already has to happen in Octavia API, because it
has to choose the worker for a F5 in the desired AZ.
That's why the worker has to communicate to the API, which AZ it's in.

Sidenote: Since this is only done once at startup, we need to guarantee that no
AZs get deleted later on.