[l2] Honor dry-run config option
When dry-run is set, the only difference is that we operate on an
empty list of BigIPs and VCMPs.
Since L2 operations are always to be applied to all BigIPs/VCMPs,
it will never be necessary for the dry-run flag to be honored anywhere
else than initialize_bigips, as long as self._bigips and self._vcmps
are using initialize_bigips.

Error reporting using all() has to be changed since all([]) returns
True. So we first need to check whether the list is empty.