[rescheduling] Fix add_loadbalancer: Don't run ensure_l2_flow if LBs already here
ensure_l2_flow is called from ControllerWorker.add_loadbalancer
unconditionally. If a load balancer of the same network (but a
different subnet) already exists on this device there will be a subnet
route (the one for the subnet of the LB-to-add) that needs cleaning up
before the SelfIP for the subnet of the LB-to-add can be
created. Also, the whole ensure_l2_flow isn't needed in this case.

So execute only sync_l2_selfips_and_subnet_routes_flow in this case.