[rescheduling] Fix: SelfIPs not removed
a1510e4 only switched out ensure_l2_flow for
sync_l2_selfips_and_subnet_routes_flow in the remove_loadbalancer
function, however one issue remained[1]:

When the LB to be removed is the last one on this device, the
shadowing loadbalancers variable is set to an empty array, which is
then supplied as argument to ensure_selfips, which in turn yields no
SelfIP ports. This means that when the else-branch is executed,
selfips will be [], so cleanup_selfips(selfips) won't delete any
ports. This is due to the loadbalancers variable being shadowed. This
error was introduced in 101fa5f with the first rescheduling
implementation.

The fix:
Instead, a new variable loadbalancers_remaining has to be created,
which contains all load balancers on this device except the one that
is to be removed.
- If there are load balancers remaining and the one to be removed was
  the last of its subnet, that subnet's SelfIP ports will be removed
  by ensure_selfips(..., cleanup_orphans=True), before the remaining
  SelfIPs (and subnet routes) are synced.
- If there are no load balancers remaining, then selfips contains only
  the SelfIP ports for the load balancer to be removed, which is then
  cleaned up with everything else L2.

[1] https://github.com/sapcc/octavia-f5-provider-driver/pull/214#pullrequestreview-1383776114