Fix: SelfIP ports not removed if LB was last of subnet
If the subnet of the LB to be removed is now empty, remove the
unneeded SelfIP ports. Since they are not considered orphaned (they
aren't even considered by ensure_selfips because the subnet is gone)
they need to be determined by comparing SelfIP ports for all LBs with
those of the remaining LBs.