fixup! Fix: When SelfIP creation fails, route is not recreated
When a route is to be replaced by a SelfIP on the F5, the route is
first deleted before the SelfIP is created. However, when SelfIP
creation fails[1], the route needs to be brought back from the
dead. This does not happen - rollback mechanisms for this are not in
place.
This commit fixes this.

[1] This can e.g. happen when a member exists on the F5 that has the
    same IP address as the SelfIP that we want to create. Since
    members can point to any arbitrary IP addresses, including
    unallocated ones, this can very easily happen.