fixup! Fix SelfIP creation error handling: Add rollback
This commit introduces sync_selfips_and_subnet_routes, which makes a
superflow over cleanup_selfips_and_subnet_routes and
ensure_selfips_and_subnet_routes, so that both will be rolled back in
the event of an exception.

This is needed for rolling back already-deleted subnet routes in the
case of a failing SelfIP creation, which can happen when a member
already exists on the BigIP and that member is pointing to the IP
address of the SelfIP that is about to be created.
So far the subnet routes (that are to be replaced by the SelfIP) were
removed, but not resurrected after SelfIP creation failed. This commit
fixes this.

Also:
- Add missing docstrings to some tasks
- Fix/reword some comments
- When an exception bubbles up to
  _do_sync_l2_selfips_and_subnet_routes_flow, log it not with print()
  but with LOG.error