fixup! L2SyncManager: Determine existing routes before SelfIP/route syncs
Flows for syncing SelfIPs and static subnet routes (ensure_l2,
remove_l2, and sync_selfips_and_subnet_routes) need to be
revertable. In order for that to be the case they need to know which
static subnet routes existed before the execute method was run. So far
the existing subnet routes were determined during execution of the
execute method, therefore the revert method wasn't able to know which
routes it has to restore. This is now fixed by introducing the
GetAllSubnetRoutesForNetwork task, which gets all static subnet routes
for a network before executing one of the aforementioned sync flows.

There might be an easier way via the store, but that is a problem for
a future refactoring, which I will do soon:tm: anyway.

Also:
- EnsureSelfIP.revert: Fix iControlREST endpoint for SelfIP deletion