[l2] Determine existing SelfIPs/Subnet routes before flow creation
Determine existing SelfIPs and subnet routes before the flow to sync
them is created. This has the advantage of making rollbacks far
simpler.

This commit is part of preparations for introducing rollback
mechanisms for SelfIP/subnet route sync.

So far, the subnet routes to create or remove are determined in the
tasks EnsureSubnetRoutes and CleanupSubnetRoutes, respectively. This
means that their revert methods would have to figure that out as
well. However, determining the prior state of subnet routes is a
difficult task after the execute method has already run, since we
don't know how many subnet routes have already been successfully
removed/created before the rollback began.

To solve this, figure out the existing and needed subnet routes at
flow creation time[1]. This allows us to create one task per subnet
route (just like EnsureSelfIP and RemoveSelfIP are already used, but
for SelfIPs) and that in turn allows their revert methods to know
exactly which subnet route to redelete/recreate.

- Replace get_selfips_from_device_for_vlan with
  get_existing_selfips_and_subnet_routes_flow

- Supply preexisting_selfips and preexisting_subnet_routes to
  get_cleanup_selfips_and_subnet_routes_flow and
  ensure_selfips_and_subnet_routes

- Rename cleanup_selfips_and_subnet_routes to
  get_cleanup_selfips_and_subnet_routes_flow

- Rename GetAllSelfIPsForVLAN to GetExistingSelfIPsForVLAN

- Add GetExistingSubnetRoutesForNetwork (used by
  get_existing_selfips_and_subnet_routes_flow)

[1] f5_flows.get_existing_selfips_and_subnet_routes_flow