[l2][unittest] Fix test_revert_RemoveSelfIP: SelfIPs to remove only contain ID
This is actually a bug in the code, not the test. To reveal the bug,
the test has to be fixed. This test is now running red, the code will
be fixed in a subsequent commit.

store['existing_selfips'] is a list of port objects, retrieved via
GetExistingSelfIPsForVLAN. However, these port objects contain only
the port ID. But when RemoveSelfIP is rolled back, it needs to know
more than the port ID, e.g. the subnet, which is retrieved from the
port's first fixed IP. Since the port only has an ID, the SelfIP
cannot be restored on the BigIP.

This commit fixes the test to reflect this and demonstrate the bug in
the code.