Temporarily roll back #298
This is needed to be able to deploy to QA without having the priority
queue feature present. We want that because we want to transport an
emergency change[1] through QA to gold while NOT transporting the
priority queue feature beyond QA yet.

After the transport has done, the priority queue feature will be added
again, together with the fix in #299

[1] https://github.com/sapcc/helm-charts/pull/8886