Fix availability check: ConnectionError exception not handled
The `requests.exceptions.Timeout` exception occurs when there is a
connection timeout or a read timeout.
But we can also get an ConnectionError exception at that point,
e. g. when the device's DNS entry is missing.

This commit patches `is_available` to catch that exception.

Also: Use early return pattern