[L2] Fix make_remove_vcmp_l2_flow: Skip interface/LAG unassignment
Skip unassigning the VLAN from an interface/LAG.  This operation was
already a NOP for iSeries. But it was implemented for rSeries. But due
to the bug on rSeries hosts that makes VLAN deletion impossible
sometimes (see the long explanation comment in f5_tasks_rseries.py),
we have to skip LAG unassignment, because in the case of this bug the
non-deleted VLAN has to stay assigned to the LAG, so that when it's
reused by a future LB, it will still have connectivity.